Seperating fuzz data producer api impl and header, using data producer on the easy fuzzers
